+++ lib/templates/app/models/organization.rb
@@ -1,0 +1,3 @@
+class Organization < ApplicationRecord
+  has_many :memberships, dependent: :restrict_with_exception
+end