+++ lib/templates/app/models/user.rb
@@ -1,0 +1,3 @@
+class User < ApplicationRecord
+  include User::Account
+end