+++ lib/templates/db/migrate/20251122141520_create_memberships.rb
@@ -1,0 +1,11 @@
+class CreateMemberships < ActiveRecord::Migration[8.0]
+  def change
+    create_table :memberships do |t|
+      t.belongs_to :user, null: false, foreign_key: true
+      t.belongs_to :organization, null: false, foreign_key: true
+      t.index [:user_id, :organization_id], unique: true
+
+      t.timestamps
+    end
+  end
+end