+++ lib/templates/lib/development/seeder.rb
@@ -1,0 +1,10 @@
+module Development
+  class Seeder
+    def self.load_seeds
+      new.load_seeds
+    end
+
+    def load_seeds
+    end
+  end
+end