+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - YARD documentation comments for all public APIs
+- Application template that sets up a new Rails application with sensible defaults
 
 ## [0.1.0] - 2025-11-25